Criando CRUD de Project file

+++ app/Http/Controllers/ProjectFileController.php
@@ -3,7 +3,7 @@
 namespace CodeProject\Http\Controllers;
 
 use CodeProject\Repositories\ProjectFileRepositoryInterface;
-use CodeProject\Services\ProjectService;
+use CodeProject\Services\ProjectFileService;
 use Illuminate\Http\Request;
 
 use CodeProject\Http\Requests;
@@ -15,7 +15,7 @@ class ProjectFileController extends Controller
     private $repository;
     private $service;
 
-    public function __construct(ProjectFileRepositoryInterface $repository, ProjectService $service)
+    public function __construct(ProjectFileRepositoryInterface $repository, ProjectFileService $service)
     {
         $this->repository = $repository;
         $this->service = $service;
@@ -25,9 +25,9 @@ class ProjectFileController extends Controller
      *
      * @return Response
      */
-    public function index()
+    public function index($id)
     {
-        //
+        return $this->repository->findWhere(['project_id'=>$id]);
     }
 
     /**
@@ -48,6 +48,18 @@ class ProjectFileController extends Controller
      */
     public function store(Request $request)
     {
+
+        $file = $request->file('file');
+        $extension = $file->getClientOriginalExtension();
+
+        $data['file'] = $file;
+        $data['extension'] = $extension;
+        $data['name'] = $request->name;
+        $data['project_id'] = $request->project_id;
+        $data['description'] = $request->description;
+
+        return $this->service->createFile($data);
+
         /* try{
          $file = $request->file('file');
          $data['file'] = $file;
@@ -65,7 +77,8 @@ class ProjectFileController extends Controller
              'message' =>  $e->getMessage()
          ];
      }*/
-    if($request->file('file'))
+
+    /*if($request->file('file'))
      {
        e = $request->file('file');
          $data['file'] = $file;
@@ -81,9 +94,25 @@ class ProjectFileController extends Controller
          'error' => true,
          'success' =>  false,
          'message' => 'Select a file to send'
-     ];
+     ];*/
 
+    }
 
+    public function showFile($id)
+    {
+        if($this->service->checkProjectPermissions($id) == false)
+        {
+            return ['error' => 'Access Forbidden'];
+        }
+
+        $filePath = $this->service->getFilePath($id);
+        $fileContent = file_get_contents($filePath);
+        $file64 = base64_encode($fileContent);
+        return [
+            'file' => $file64,
+            'size' => filesize($filePath),
+            'name' => $this->service->getFileName($id)
+        ];
     }
 
     /**
@@ -94,7 +123,11 @@ class ProjectFileController extends Controller
      */
     public function show($id)
     {
-        //
+        if($this->service->checkProjectPermissions($id) == false)
+        {
+            return ['error' => 'Access Forbidden'];
+        }
+        return $this->repository->find($id);
     }
 
     /**
@@ -117,7 +150,11 @@ class ProjectFileController extends Controller
      */
     public function update(Request $request, $id)
     {
-        //
+        if($this->service->checkProjectPermissions($id) == false)
+        {
+            return ['error' => 'Access Forbidden'];
+        }
+        return $this->service->update($request->all(), $id);
     }
 
     /**
@@ -128,6 +165,10 @@ class ProjectFileController extends Controller
      */
     public function destroy($id)
     {
+        if($this->service->checkProjectPermissions($id) == false)
+        {
+            return ['error' => 'Access Forbidden'];
+        }
         return $this->service->removeFile($id);
     }
 }
